nuclei-viewer/backend/requirements.txt
# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Data processing (use loose versions for packages in main requirements.txt)
zarr  # Already specified in main requirements.txt
numcodecs<0.16
dask  # Already specified in main requirements.txt with specific version
numpy<2.0  # Already constrained in main requirements.txt
scipy>=1.11.0
scikit-image>=0.22.0
pandas>=2.1.0

# Image processing
tifffile>=2023.9.0
Pillow>=10.1.0

# Configuration and validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml  # Already specified in main requirements.txt

# CORS and middleware
python-multipart>=0.0.6

# Logging and utilities
python-json-logger>=2.0.7